Split file of src/utils/pluralize
